[TEST] Consolidate fragmented test files into test_diff2typo.py and test_typostats.py
Consolidates fragmented, tiny, and redundant test files (test_diff2typo_alignment.py, test_typostats_coverage_gap.py, test_typostats_directory.py, test_typostats_interactive.py, and test_typostats_main.py) into their respective core domains: test_diff2typo.py and test_typostats.py.
Refactors fragile tests into clean, atomic unit tests without console prints or comments. Removes the redundant files to improve test hygiene, discoverability, and maintainability.
